Revert "D3 profile"
