+++ components/Newsletter/index.js
@@ -127,7 +127,6 @@ const CustomForm = ({
 				ba
 				bw1
 				b__blue
-				pointer
 				type="submit"
 			>
 				Subscribe
