--- components/UI/Tachyons/index.js
@@ -1,30 +1,0 @@
-import styled from "styled-components";
-import tachyons from "styled-components-tachyons";
-
-export const H4 = styled.h4`
-	${tachyons}
-`;
-export const H3 = styled.h3`
-	${tachyons}
-`;
-export const H2 = styled.h2`
-	${tachyons}
-`;
-export const H1 = styled.h1`
-	${tachyons}
-`;
-export const P = styled.p`
-	${tachyons}
-`;
-export const Img = styled.img`
-	${tachyons}
-`;
-export const A = styled.a`
-	${tachyons}
-`;
-export const Button = styled.button`
-	${tachyons}
-`;
-export const Div = styled.div`
-	${tachyons}
-`;
