--- lib/awards.js
@@ -1,38 +1,0 @@
-export default [
-	{
-		name: "1,000km",
-		distance: 1000,
-		roundal:
-			"//images.ctfassets.net/6hyijb95boju/g2Qjba48pGADx0zwAQKlo/5eb7bdf2ba688c493430c462684ff664/DotWatcher-No-To-CO2-Award_background.png"
-	},
-	{
-		name: "2,000km",
-		distance: 2000,
-		roundal:
-			"//images.ctfassets.net/6hyijb95boju/g2Qjba48pGADx0zwAQKlo/5eb7bdf2ba688c493430c462684ff664/DotWatcher-No-To-CO2-Award_background.png"
-	},
-	{
-		name: "3,000km",
-		distance: 3000,
-		roundal:
-			"//images.ctfassets.net/6hyijb95boju/g2Qjba48pGADx0zwAQKlo/5eb7bdf2ba688c493430c462684ff664/DotWatcher-No-To-CO2-Award_background.png"
-	},
-	{
-		name: "5,000km",
-		distance: 5000,
-		roundal:
-			"//images.ctfassets.net/6hyijb95boju/g2Qjba48pGADx0zwAQKlo/5eb7bdf2ba688c493430c462684ff664/DotWatcher-No-To-CO2-Award_background.png"
-	},
-	{
-		name: "10,000km",
-		distance: 10000,
-		roundal:
-			"//images.ctfassets.net/6hyijb95boju/g2Qjba48pGADx0zwAQKlo/5eb7bdf2ba688c493430c462684ff664/DotWatcher-No-To-CO2-Award_background.png"
-	},
-	{
-		name: "12,000km",
-		distance: 12000,
-		roundal:
-			"//images.ctfassets.net/6hyijb95boju/g2Qjba48pGADx0zwAQKlo/5eb7bdf2ba688c493430c462684ff664/DotWatcher-No-To-CO2-Award_background.png"
-	}
-];
