+++ lib/user.js
