+++ now.json
@@ -60,6 +60,10 @@
 			"dest": "/profile/edit"
 		},
 		{
+			"src": "/profile/(?<name>[^/]+)$",
+			"dest": "/profile?name=$name"
+		},
+		{
 			"src": "/calendar/(?<year>[^/]+)?/(?<month>[^/]+)$",
 			"dest": "/calendar?year=$year&month=$month"
 		}
