+++ package.json
@@ -31,8 +31,6 @@
 		"classnames": "^2.2.6",
 		"cloudinary-react": "^1.3.0",
 		"contentful": "^7.10.0",
-		"d3": "^5.16.0",
-		"d3-legend": "^1.0.0",
 		"date-fns": "^1.30.1",
 		"disqus-react": "^1.0.7",
 		"formik": "^2.1.2",
@@ -69,7 +67,6 @@
 		"remark-breaks": "^1.0.3",
 		"remark-shortcodes": "^0.2.1",
 		"slugify": "^1.3.5",
-		"smoothscroll-polyfill": "^0.4.4",
 		"styled-components": "^5.0.1",
 		"styled-components-tachyons": "^0.0.5",
 		"tachyons": "^4.11.1",
