+++ pages/_document.js
@@ -4,15 +4,13 @@ import React from "react";
 
 import * as Sentry from "@sentry/browser";
 
-if (process.env.NODE_ENV === "production") {
-	process.on("unhandledRejection", err => {
-		Sentry.captureException(err);
-	});
+process.on("unhandledRejection", err => {
+	Sentry.captureException(err);
+});
 
-	process.on("uncaughtException", err => {
-		Sentry.captureException(err);
-	});
-}
+process.on("uncaughtException", err => {
+	Sentry.captureException(err);
+});
 
 export default class Document extends NextDocument {
 	static getInitialProps({ renderPage }) {
